more tests and cleanup
Moved the fetchGroups test to provider_azure_test.go, added another
unit test for getClaimSources().
